fix(flashcards): split flashcard-embed into server-safe + client files (Epoch 19, Sprint 4)
The original Sprint 4 commit kept the FlashcardEmbed client Node and the
ServerFlashcardEmbed server Node in the same file. That file imported
`react-dom/client` at module top-level for the createRoot mount bridge.
Next.js 16 (Turbopack) rejected this the moment any server-side import
chain reached the file:

  ./lib/domain/editor/extensions/blocks/flashcard-embed.tsx
  ← ./lib/domain/editor/extensions-server.ts
  ← ./lib/domain/content/markdown.ts
  ← ./extensions/workplaces/server/service.ts
  ← ./app/api/content/workspaces/route.ts

Error: "You're importing a component that imports react-dom/client. It
only works in a Client Component but none of its parents are marked
with 'use client'."

Fix
  Split into two files. extensions-server.ts and collaboration/
  extensions.ts continue to import ServerFlashcardEmbed from the
  unchanged path; extensions-client.ts now imports FlashcardEmbed from
  the new sibling file.

  lib/domain/editor/extensions/blocks/flashcard-embed.tsx (server-safe)
    - Block schema via createBlockSchema
    - registerBlock() call (registry is a plain Map; server-safe)
    - flashcardEmbedAttrSpec() factory shared by both Nodes so client +
      server can never drift on the schema
    - ServerFlashcardEmbed Node.create
    - Zero React / DOM imports

  lib/domain/editor/extensions/blocks/flashcard-embed-client.tsx (NEW)
    - Imports react / react-dom/client / FlashcardEmbedNodeView
    - renderFlashcardEmbed() bridge that mounts the React tree via
      createRoot, stores root + cleanup on contentDom
    - FlashcardEmbed Node.create with addNodeView using createBlockNodeView

  extensions-client.ts: import path -> ./extensions/blocks/flashcard-embed-client

The precedent blocks (excalidraw-block, periodic-summary) didn't hit
this issue because they render via vanilla DOM. The React-mount approach
needs the file split — same pattern as any other component that needs
Next.js client/server isolation.

Verified
  - tsc --noEmit exit 0
  - eslint --max-warnings 159 holds at 159
