+++ lib/domain/editor/extensions-client.ts
@@ -65,7 +65,7 @@ import { HabitTracker } from "./extensions/blocks/habit-tracker";
 import { Stopwatch } from "./extensions/blocks/stopwatch";
 import { PullQuote } from "./extensions/blocks/pull-quote";
 import { TableOfContents } from "./extensions/blocks/table-of-contents";
-import { FlashcardEmbed } from "./extensions/blocks/flashcard-embed";
+import { FlashcardEmbed } from "./extensions/blocks/flashcard-embed-client";
 import {
   UnsupportedBlock,
   UnsupportedInline,
